rename: CafegoryToken 클래스 이름을 JwtToken으로 변경 및 token 패키지로 이동
