+++ src/main/java/com/example/demo/implement/token/JwtToken.java
@@ -1,11 +1,11 @@
-package com.example.demo.dto.auth;
+package com.example.demo.implement.token;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
 @Getter
 @RequiredArgsConstructor
-public class CafegoryToken {
+public class JwtToken {
 
 	private final String accessToken;
 	private final String refreshToken;
