+++ src/test/java/com/example/demo/apidocs/OAuthControllerApiTest.java
@@ -2,7 +2,7 @@ package com.example.demo.apidocs;
 
 import com.epages.restdocs.apispec.ResourceSnippetParameters;
 import com.example.demo.controller.OAuthController;
-import com.example.demo.dto.auth.CafegoryToken;
+import com.example.demo.implement.token.JwtToken;
 import com.example.demo.service.facade.OAuthFacade;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -53,7 +53,7 @@ public class OAuthControllerApiTest {
 
     @Test
     void kakao() throws Exception {
-        when(oAuthFacade.handleOauthLogin(any())).thenReturn(new CafegoryToken("access-token-value", "refresh-token-value"));
+        when(oAuthFacade.handleOauthLogin(any())).thenReturn(new JwtToken("access-token-value", "refresh-token-value"));
 
         this.mockMvc.perform(
                 get("/oauth2/kakao?code={code}", "eyJ0eXAiOiJKV1QiLCJhbGciOiJIU")
